refactor(agents): share agent entry and block reply payload types
