+++ src/agents/pi-embedded-payloads.ts
@@ -1,0 +1,8 @@
+export type BlockReplyPayload = {
+  text?: string;
+  mediaUrls?: string[];
+  audioAsVoice?: boolean;
+  replyToId?: string;
+  replyToTag?: boolean;
+  replyToCurrent?: boolean;
+};
